requirements-dev.txt
pytest==6.0.1
mypy==0.770

flake8==3.8.3
coverage==5.3
black==20.8b1

django-extensions==3.0.9
django-debug-toolbar==3.1.1